Método melhorado e mudança de nome do job
